requirements.txt
eyes-selenium==5.26.0
Appium-Python-Client==1.3.0
pytest==7.3.1
urllib3==1.26.19
